requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
matplotlib>=3.5.0
seaborn>=0.11.0
mysql-connector-python>=8.0.0
numpy>=1.21.0
plotly>=5.0.0
